Divide Footer component to sub components for better reability and handling
