+++ src/components/footer/index.js
@@ -1,11 +1,5 @@
 import React, { PureComponent } from 'react';
-import Tooltip from '@tippy.js/react';
-import Lottie from 'react-lottie';
-import {
-  Transition,
-  animated,
-  interpolate
-} from 'react-spring/renderprops.cjs';
+import { Transition, animated } from 'react-spring/renderprops.cjs';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 import isEqual from 'lodash/isEqual';
@@ -16,22 +10,19 @@ import {
   getPasswordModalStatus,
   getComponentWarnings
 } from './../../reducers';
-import animData from './../../assets/anims/wave.json';
+
+import DotAnim from '../DotAnim';
+import LoadingIndicator from './LoadingIndicator';
+import ModuleStatus from './ModuleStatus';
+import ProviderStatus from './ProviderStatus';
+import SocialBar from './SocialBar';
+import StatusMessage from './StatusMessage';
+import Wave from './Wave';
 
 import { componentStatus } from './../../constants/statusDicts';
-import golem_loading from './../../assets/img/golem-loading.svg';
 
 const { remote, ipcRenderer } = window.electron;
 const currentPlatform = remote.getGlobal('process').platform;
-
-const defaultOptions = {
-  loop: true,
-  autoplay: true,
-  animationData: animData,
-  rendererSettings: {
-    preserveAspectRatio: 'xMidYMid slice'
-  }
-};
 
 const ISSUES = {
   PORT: {
@@ -74,6 +65,19 @@ function isGolemConnecting(isEngineOn, status) {
         status.client.status !== componentStatus.SHUTDOWN)) &&
     !status.client.message.includes('configuration')
   );
+}
+
+function golemDotClass(status, connectionProblem, componentWarnings = []) {
+  if (status && isGolemConnected(status)) {
+    return connectionProblem?.status || componentWarnings.length > 0
+      ? componentWarnings.length === 1 && componentWarnings[0].issue === 'RAM'
+        ? 'blue'
+        : 'yellow'
+      : 'green';
+  } else if (status?.status !== componentStatus.EXCEPTION) {
+    return 'yellow';
+  }
+  return 'red';
 }
 
 const mapStateToProps = state => ({
@@ -135,157 +139,9 @@ export class Footer extends PureComponent {
     }
   };
 
-  //TODO re-write it cleaner
-  golemDotClass(status, connectionProblem, componentWarnings = []) {
-    if (status && isGolemConnected(status)) {
-      return connectionProblem?.status || componentWarnings.length > 0
-        ? componentWarnings.length === 1 && componentWarnings[0].issue === 'RAM'
-          ? 'blue'
-          : 'yellow'
-        : 'green';
-    } else if (status?.status !== componentStatus.EXCEPTION) {
-      return 'yellow';
-    }
-    return 'red';
-  }
-
-  _loadErrorUrl = msg => {
-    switch (msg) {
-      case 'Error creating Docker VM': //docker
-        return (
-          <a
-            href={
-              currentPlatform === 'win32'
-                ? 'https://docs.golem.network/#/Products/Brass-Beta/Issues-&-Troubleshooting?id=docker-errors-on-windows-10'
-                : 'https://docs.golem.network/#/Products/Brass-Beta/Issues-&-Troubleshooting?id=docker-errors-on-macos'
-            }>
-            <span className="icon-new-window" />
-          </a>
-        );
-      case 'Outdated hyperg version': //hyperg
-        return (
-          <a href="https://docs.golem.network/#/Products/Brass-Beta/Issues-&-Troubleshooting?id=outdated-hyperg-version">
-            <span className="icon-new-window" />
-          </a>
-        );
-      case 'Chain sync error': //sync
-        return (
-          <a href="https://docs.golem.network/#/Products/Brass-Beta/Issues-&-Troubleshooting?id=sync">
-            <span className="icon-new-window" />
-          </a>
-        );
-        break;
-      case 'Error connecting geth': //geth
-        return (
-          <a href="https://docs.golem.network/#/Products/Brass-Beta/Issues-&-Troubleshooting?id=geth">
-            <span className="icon-new-window" />
-          </a>
-        );
-      default:
-        break;
-    }
-  };
-
-  _openLogs = () => {
-    ipcRenderer.send('open-logs');
-  };
-
   _cancelShutdown = () => this.props.actions.gracefulShutdown();
 
-  _forceQuit = () => {
-    this.props.actions.toggleForceQuit();
-  };
-
-  _fetchState(stat) {
-    if (stat) {
-      let state = stat.status;
-      if (stat?.environment) {
-        state += this._fetchEnvironment(stat.environment);
-      }
-      return state;
-    }
-  }
-
-  _fetchEnvironment(env) {
-    switch (env) {
-      case 'BLENDER':
-        return ' (CPU - Blender)';
-      case 'BLENDER_NVGPU':
-        return ' (GPU - Blender)';
-      case 'BLENDER_SGX':
-        return ' (SGX - Blender)';
-      case 'WASM':
-        return ' (CPU - gWasm)';
-      default:
-        return '';
-    }
-  }
-
-  _loadConnectionWarnings(status, connectionProblem, componentWarnings = []) {
-    let warningMessage = '';
-    const newLineBeforeWarning =
-      status?.client?.message.length > 10 ? <br key="br" /> : ' ';
-
-    if (connectionProblem.status)
-      warningMessage =
-        connectionProblem.issue == 'WEBSOCKET' ? (
-          <span key="warningWebsocket" className="info__warnings">
-            connection dropped
-          </span>
-        ) : (
-          ' '
-        );
-    else if (componentWarnings.length > 0) {
-      warningMessage = (
-        <span key="warningComponent" className="info__warnings">
-          {componentWarnings.length > 1
-            ? `${componentWarnings.length} issues`
-            : componentWarnings[0].status &&
-              ISSUES[componentWarnings[0].issue].title}
-          <Tooltip
-            interactive
-            className="tooltip__warning-component"
-            content={
-              <p className="info__connection">
-                {componentWarnings.map((item, index) => {
-                  const { docs, message } = ISSUES[
-                    componentWarnings[index].issue
-                  ];
-                  return (
-                    <span
-                      key={index.toString()}
-                      className="info__connection__item">
-                      <span
-                        className={`icon-status-dot ${
-                          componentWarnings[index].issue === 'RAM'
-                            ? 'icon-status-dot--info'
-                            : ''
-                        }`}
-                      />
-                      {message}
-                      {componentWarnings[index].issue === 'RAM' &&
-                        componentWarnings[index]?.value &&
-                        `${componentWarnings[index]?.value} GiB`}
-                      <a href={docs}>
-                        <span className="icon-new-window" />
-                      </a>
-                    </span>
-                  );
-                })}
-              </p>
-            }
-            distance={status?.client?.message.length > 10 ? 40 : 30}
-            placement="top"
-            trigger="mouseenter"
-            theme="light">
-            <span className="icon-warning-rounded" />
-          </Tooltip>
-        </span>
-      );
-    }
-
-    return [newLineBeforeWarning, warningMessage];
-  }
+  _forceQuit = () => this.props.actions.toggleForceQuit();
 
   render() {
     const {
@@ -300,9 +156,6 @@ export class Footer extends PureComponent {
       passwordModal,
       version
     } = this.props;
-    const versionTemplate = version?.error
-      ? version?.message || ''
-      : `${version?.message || ''}${version?.number || ''}`;
     return (
       <div
         className={`content__footer-main ${isGolemConnecting(
@@ -312,65 +165,20 @@ export class Footer extends PureComponent {
         <div className="section__actions">
           <div className="section__actions-status">
             <span
-              className={`progress-status indicator-status indicator-status--${this.golemDotClass(
+              className={`progress-status indicator-status indicator-status--${golemDotClass(
                 status.client,
                 connectionProblem,
                 componentWarnings
               )}`}
             />
             <div>
-              <span>
-                <span className="status-message">
-                  <span>
-                    {status?.client?.message ? (
-                      isGolemConnecting(isEngineOn, status) ? (
-                        <span>
-                          {status.client.message}
-                          <Tooltip
-                            content={
-                              <p className="info__connection">
-                                The process may take a few seconds.
-                                <br />
-                                When all connection statuses are green
-                                <br />
-                                then app will properly connect.
-                              </p>
-                            }
-                            placement="top"
-                            trigger="mouseenter">
-                            <span className="icon-question-mark" />
-                          </Tooltip>
-                        </span>
-                      ) : (
-                        status.client.message
-                      )
-                    ) : (
-                      <span>
-                        Loading
-                        <span className="jumping-dots">
-                          <span className="dot-1">.</span>
-                          <span className="dot-2">.</span>
-                          <span className="dot-3">.</span>
-                        </span>
-                      </span>
-                    )}
-                  </span>
-                  {status && status[0] && (
-                    <span>
-                      <a href="https://docs.golem.network/#/Products/Brass-Beta/Installation">
-                        <span className="icon-new-window" />
-                      </a>
-                    </span>
-                  )}
-                </span>
-                {status?.client?.message &&
-                  this._loadErrorUrl(status.client.message)}
-                {this._loadConnectionWarnings(
-                  status,
-                  connectionProblem,
-                  componentWarnings
-                )}
-              </span>
+              <StatusMessage
+                componentWarnings={componentWarnings}
+                connectionProblem={connectionProblem}
+                isEngineOn={isEngineOn}
+                isGolemConnecting={isGolemConnecting}
+                status={status}
+              />
               <Transition
                 native
                 initial={null}
@@ -397,130 +205,21 @@ export class Footer extends PureComponent {
                 {toggle =>
                   toggle
                     ? props => (
-                        <animated.div
-                          style={{
-                            opacity: props.opacity.interpolate(
-                              opacity => opacity
-                            ),
-                            transform: props.transform.interpolate(
-                              y => `translateX(${y}px)`
-                            ),
-                            position: props.position
-                          }}
-                          className="status-node">
-                          <span>
-                            Provider state:{' '}
-                            {this._fetchState(stats.provider_state)}
-                          </span>
-                          <br />
-                          {status?.client?.status ===
-                            componentStatus.SHUTDOWN ||
-                          isGracefulShutdownEnabled ? (
-                            <div className="action__graceful-shutdown">
-                              <div
-                                className="action__graceful-shutdown-item"
-                                onClick={this._cancelShutdown}>
-                                <span className="icon-failure" />
-                                <span>Cancel shutdown</span>
-                              </div>
-                              <div
-                                className="action__graceful-shutdown-item"
-                                onClick={this._forceQuit}>
-                                <span className="icon-force-quit" />
-                                <span>Force quit</span>
-                              </div>
-                            </div>
-                          ) : (
-                            [
-                              <span key="stats_01">
-                                Attempted:{' '}
-                                {stats.subtasks_computed &&
-                                  stats.subtasks_computed[1] +
-                                    stats.subtasks_with_timeout[1] +
-                                    stats.subtasks_with_errors[1]}
-                              </span>,
-                              <br key="stats_02" />,
-                              <span key="stats_03">
-                                {stats.subtasks_with_errors &&
-                                  `${
-                                    stats.subtasks_with_errors[1]
-                                  } error | ${stats.subtasks_with_timeout &&
-                                    stats
-                                      .subtasks_with_timeout[1]} timeout | ${stats.subtasks_accepted &&
-                                    stats.subtasks_accepted[1]} success`}
-                              </span>
-                            ]
-                          )}
-                        </animated.div>
+                        <ProviderStatus
+                          {...props}
+                          cancelShutdown={this._cancelShutdown}
+                          forceQuit={this._forceQuit}
+                          stats={stats}
+                          status={status}
+                        />
                       )
                     : props => (
-                        <animated.div
-                          style={{
-                            opacity: props.opacity.interpolate(
-                              opacity => opacity
-                            ),
-                            transform: props.transform.interpolate(
-                              y => `translateX(${y}px)`
-                            ),
-                            position: props.position
-                          }}
-                          className="status-node__loading">
-                          {status?.client?.status &&
-                          status.client.status !== componentStatus.EXCEPTION ? (
-                            <div className="status__components">
-                              <div className="item__status">
-                                <div>
-                                  <span
-                                    className={`component-dot component-dot--${this.golemDotClass(
-                                      status?.hyperdrive,
-                                      connectionProblem
-                                    )}`}
-                                  />
-                                  <span>Hyperg: </span>
-                                </div>
-                                <span>{status?.hyperdrive?.message}</span>
-                              </div>
-                              <div className="item__status">
-                                <div>
-                                  <span
-                                    className={`component-dot component-dot--${this.golemDotClass(
-                                      status?.hypervisor,
-                                      connectionProblem
-                                    )}`}
-                                  />
-                                  <span>Hypervisor: </span>
-                                </div>
-                                <span>{status?.hypervisor?.message}</span>
-                              </div>
-                              <div className="item__status">
-                                <div>
-                                  <span
-                                    className={`component-dot component-dot--${this.golemDotClass(
-                                      status?.docker,
-                                      connectionProblem
-                                    )}`}
-                                  />
-                                  <span>Docker: </span>
-                                </div>
-                                <span>{status?.docker?.message}</span>
-                              </div>
-                              <div className="item__status">
-                                <div>
-                                  <span
-                                    className={`component-dot component-dot--${this.golemDotClass(
-                                      status?.ethereum,
-                                      connectionProblem
-                                    )}`}
-                                  />
-                                  <span>Geth: </span>
-                                </div>
-                                <span>{status?.ethereum?.message}</span>
-                              </div>
-                            </div>
-                          ) : (
-                            <span>Error while fetching status</span>
-                          )}
-                        </animated.div>
+                        <ModuleStatus
+                          {...props}
+                          connectionProblem={connectionProblem}
+                          golemDotClass={golemDotClass}
+                          status={status}
+                        />
                       )
                 }
               </Transition>
@@ -532,42 +231,10 @@ export class Footer extends PureComponent {
             disabled={isGolemConnecting(isEngineOn, status)}>
             {isEngineOn ? 'Stop' : 'Start'} Golem
           </button>
-          {
-            <div className="wave-loading" id="waveLoading">
-              <Lottie
-                width={'100%'}
-                options={defaultOptions}
-                isStopped={this.state.stopAnim}
-              />
-            </div>
-          }
+          <Wave stopAnim={this.state.stopAnim} />
         </div>
-        <div className="content__footer-social">
-          <span className="element__footer" onClick={this._openLogs}>
-            <span className="icon-logs" />
-            <u>open logs</u>
-          </span>
-          <a
-            className="element__footer"
-            href="https://www.github.com/golemfactory">
-            <span className="icon-golem" />
-            {versionTemplate}
-          </a>
-          <a className="element__footer" href="https://chat.golem.network">
-            <span className="icon-chat" />
-            <u>golem chat</u>
-          </a>
-        </div>
-        <div>
-          <div
-            className={`loading-indicator ${isEngineLoading ? 'active' : ''}`}
-          />
-          <object
-            className={`loading-icon ${isEngineLoading ? 'active' : ''}`}
-            type="image/svg+xml"
-            data={golem_loading}
-          />
-        </div>
+        <SocialBar version={version} />
+        <LoadingIndicator isEngineLoading={isEngineLoading} />
       </div>
     );
   }
