+++ src/scss/footer.scss
@@ -228,7 +228,6 @@
     .loading-indicator {
         position: absolute;
         z-index: 9 !important;
-        position: absolute;
         bottom: 40px;
         right: 90px;
         width: 0;
@@ -249,18 +248,14 @@
 
     .loading-icon {
         position: absolute;
-        display: none;
+        display: block;
         left: 0;
         right: 0;
-        top: 0;
+        top: -70vh;
         bottom: 0;
         margin: auto;
         width: 120px;
         z-index: 11;
-
-        &.active {
-            display: block;
-        }
     }
 }
